% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_about.R
\name{aboutServer}
\alias{aboutServer}
\title{Server Module for About panel}
\usage{
aboutServer(id)
}
\arguments{
\item{id}{The namespace ID for the module.}
}
\description{
Server Module for About panel
}
