% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_about.R
\name{aboutUI}
\alias{aboutUI}
\title{UI Module for About panel}
\usage{
aboutUI(id)
}
\arguments{
\item{id}{The namespace ID for the module.}
}
\description{
This function initializes and runs the Shiny application.
}
