% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app_server}
\alias{app_server}
\title{Main server function for the application}
\usage{
app_server(input, output, session, preloaded_data = NULL)
}
\arguments{
\item{input}{Input objects}

\item{output}{Output objects}

\item{session}{Session object}

\item{preloaded_data}{Preloaded overview table data}
}
\description{
Main server function for the application
}
