% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oddjob.R
\name{call_oddjob}
\alias{call_oddjob}
\title{Call Oddjob via the GitLab API}
\usage{
call_oddjob(table = NULL)
}
\arguments{
\item{table}{Optional string to specify the table name.}
}
\value{
Results from the GitLab API.
}
\description{
Call Oddjob via the GitLab API
}
