% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_module.R
\name{dbServer}
\alias{dbServer}
\title{UI Module for Summary panel}
\usage{
dbServer(id, db_config = NULL)
}
\arguments{
\item{id}{ID for the module}

\item{db_config}{Configuration list for database connection}
}
\description{
UI Module for Summary panel
}
