% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckDB.R
\name{duckdb_table}
\alias{duckdb_table}
\title{Get a DuckDB table}
\usage{
duckdb_table(table = c("pointers", "columns", "global_data"))
}
\arguments{
\item{table}{Table name}
}
\value{
A data frame containing the requested table.
}
\description{
Get a DuckDB table
}
