% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oddjob.R
\name{get_oddjob}
\alias{get_oddjob}
\title{Get Information from Oddjob}
\usage{
get_oddjob(table_name = NULL)
}
\arguments{
\item{table_name}{Name of the table to retrieve information for.}
}
\value{
List with meta data.
}
\description{
Get Information from Oddjob
}
