% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_module.R
\name{inputUI}
\alias{inputUI}
\title{Common input module for selecting datasets and versions}
\usage{
inputUI(id)
}
\arguments{
\item{id}{ID}
}
\description{
Common input module for selecting datasets and versions
}
