% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_labelmatrix.R
\name{labelMatrixUI}
\alias{labelMatrixUI}
\title{Label matrix user interface}
\usage{
labelMatrixUI(id)
}
\arguments{
\item{id}{The namespace ID for the module.}
}
\description{
Label matrix user interface
}
