% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_TypeMatrixWeb}
\alias{plot_TypeMatrixWeb}
\title{Plot the Type Matrix}
\usage{
plot_TypeMatrixWeb(table, skip = 0, clip_date = TRUE)
}
\arguments{
\item{table}{Table/Data Name}

\item{skip}{Skip the first n versions}

\item{clip_date}{Logical, if TRUE, it will remove the time part from the version}
}
\value{
A ggplot object
}
\description{
Plot the Type Matrix
}
