% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_presencematrix.R
\name{presenceMatrixUI}
\alias{presenceMatrixUI}
\title{UI Module for Presence Matrix panel}
\usage{
presenceMatrixUI(id)
}
\arguments{
\item{id}{ID for the module}
}
\description{
UI Module for Presence Matrix panel
}
