% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_scatterplots.R
\name{scatterplotModuleServer}
\alias{scatterplotModuleServer}
\title{Server function for scatterplot module}
\usage{
scatterplotModuleServer(id)
}
\arguments{
\item{id}{The module ID}
}
\value{
A Shiny module server function
}
\description{
Server function for scatterplot module
}
