% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_scatterplots.R
\name{scatterplotModuleUI}
\alias{scatterplotModuleUI}
\title{UI function for scatterplot module}
\usage{
scatterplotModuleUI(id)
}
\arguments{
\item{id}{The module ID}
}
\value{
A shiny UI element with control inputs
}
\description{
UI function for scatterplot module
}
