% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_scatterplots.R
\name{scatterplotOutputUI}
\alias{scatterplotOutputUI}
\title{UI function for scatterplot module output}
\usage{
scatterplotOutputUI(id)
}
\arguments{
\item{id}{The module ID}
}
\value{
A shiny UI element with the plot output
}
\description{
UI function for scatterplot module output
}
