% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_summary.R
\name{summaryServer}
\alias{summaryServer}
\title{Server Module for Summary panel}
\usage{
summaryServer(id, dataset)
}
\arguments{
\item{id}{ID for the module}

\item{dataset}{Reactive expression that returns the dataset to be summarized}
}
\description{
Server Module for Summary panel
}
