% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table_overview}
\alias{table_overview}
\title{Table Validation Overview}
\usage{
table_overview(data)
}
\arguments{
\item{data}{Data File}
}
\value{
A \code{gt} table with the validation overview.
}
\description{
Table Validation Overview
}
