% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table_pointer}
\alias{table_pointer}
\title{Create a table with pointer information}
\usage{
table_pointer(pointer_name = "penguins", date = "2025-08-20T13-52-15")
}
\arguments{
\item{pointer_name}{Name of the pointer/data}
}
\value{
A gt table with pointer information
}
\description{
Create a table with pointer information
}
