% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_octopussy.R
\name{validationReportServer}
\alias{validationReportServer}
\title{Validation report server logic}
\usage{
validationReportServer(id, dataset, version)
}
\arguments{
\item{id}{ID for the module}

\item{dataset}{Reactive expression for the selected dataset}

\item{version}{Reactive expression for the selected version}
}
\description{
Validation report server logic
}
