% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_octopussy.R
\name{validationReportUI}
\alias{validationReportUI}
\title{Validation user interface}
\usage{
validationReportUI(id)
}
\arguments{
\item{id}{ID for the module}
}
\description{
Validation user interface
}
