% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_labelmatrix.R
\name{labelMatrixServer}
\alias{labelMatrixServer}
\title{Server Module for Type Matrix panel}
\usage{
labelMatrixServer(id, dataset)
}
\arguments{
\item{id}{The namespace ID for the module.}

\item{dataset}{A reactive expression that returns the selected dataset.}
}
\description{
Server Module for Type Matrix panel
}
