% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_labelmatrix.R
\name{labelMatrixUI}
\alias{labelMatrixUI}
\title{UI Module for labelMatrix}
\usage{
labelMatrixUI(id)
}
\arguments{
\item{id}{The namespace ID for the module.}
}
\description{
UI Module for labelMatrix
}
