% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_presencematrix.R
\name{presenceMatrixServer}
\alias{presenceMatrixServer}
\title{Server Module for Presence Matrix panel}
\usage{
presenceMatrixServer(id, dataset)
}
\arguments{
\item{id}{ID for the module}

\item{dataset}{Reactive expression that returns the dataset to be visualized}
}
\description{
Server Module for Presence Matrix panel
}
