% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_typematrix.R
\name{typeMatrixServer}
\alias{typeMatrixServer}
\title{Server Module for Type Matrix panel}
\usage{
typeMatrixServer(id, dataset)
}
\arguments{
\item{id}{ID for the module}

\item{dataset}{Reactive expression that returns the selected dataset}
}
\description{
Server Module for Type Matrix panel
}
