% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_octopussy.R
\name{validationReportServer}
\alias{validationReportServer}
\title{Server Module for Validation Report panel}
\usage{
validationReportServer(id, dataset)
}
\description{
Server Module for Validation Report panel
}
