% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_octopussy.R
\name{validationReportUI}
\alias{validationReportUI}
\title{UI Module for Validation Report}
\usage{
validationReportUI(id)
}
\arguments{
\item{id}{ID for the module}
}
\description{
UI Module for Validation Report
}
